Added more to Drive code
moved some constants related to drive code and controller tolerances from main code to here. FUTURE: add a constructor for passing in an array of pins for the motors. 2 member functions added setStickPwr, basically just moved code from main to here. and generateTurnScalar, need to work with Grant on this to fully implement